mods to entry script
Change-type: patch

+++ entrypoint.sh
@@ -14,24 +14,27 @@ function get_version () {
 }
 
 # Ensure we have a repo.yml file
-if [ ! -f repo.yml ]; then
-  echo "repo.yml not found, creating generic..."
+if [[ ! -f package.json && ! -f repo.yml ]]; then
+  echo "No package.json or repo.yml found, creating generic type project..."
   echo "type: generic" > repo.yml
 fi
 
-git tag
 get_version
-echo $VERSION
+echo "Current version: "$VERSION
 
 # Run versionist
 balena-versionist
 
 get_version
+echo "New version:"$VERSION
 
-# Commit and push changes
-git config --local user.email "$INPUT_GITHUB_EMAIL"
-git config --local user.name "$INPUT_GITHUB_USERNAME"
-git add .
-git commit -m "$VERSION"
-git tag -a "$VERSION" -m "$VERSION"
-git push "https://${GITHUB_ACTOR}:${INPUT_GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git" HEAD:master --follow-tags
+# # Commit and push changes
+# git config --local user.email "$INPUT_GITHUB_EMAIL"
+# git config --local user.name "$INPUT_GITHUB_USERNAME"
+# git add .
+# git commit -m "$VERSION"
+# git tag -a "$VERSION" -m "$VERSION"
+# git push "https://${GITHUB_ACTOR}:${INPUT_GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git" HEAD:master --follow-tags
+
+# Push to NPM if a token was provided
+# npm publish
